backend/requirements.txt
Django~=4.2.9
psycopg2-binary~=2.9.9
django-environ~=0.9.0
django-cors-headers
PyJWT